pyproject.toml
[project]
name = "amplifier-module-provider-gemini"
version = "1.0.0"
description = "Reference provider module - Google Gemini integration for Amplifier"
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Microsoft MADE:Explorations Team" },
]
dependencies = [
    "google-genai>=1.40.0",
]

[project.entry-points."amplifier.modules"]
provider-gemini = "amplifier_module_provider_gemini:mount"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = [
    "amplifier_module_provider_gemini",
]

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
